Move reactor context propagation operator instrumentation to separate instrumentation module (#5188)
